commit before build problem

+++ components/comments/DeletePostButton.jsx
@@ -4,20 +4,54 @@ import { deleteComment } from "@/app/actions/commentActions";
 import { Loader } from "@/components/ui/loader";
 import { Trash } from "lucide-react";
 import { useTransition } from "react";
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from "../ui/alert-dialog";
+import { Button } from "../ui/button";
 
 export const DeletePostButton = ({ commentId }) => {
   const [isPending, starTransaction] = useTransition();
 
   return (
-    <button
-      className="flex items-center gap-1 rounded-md hover:bg-accent"
-      onClick={() => {
-        starTransaction(() => {
-          deleteComment(commentId);
-        });
-      }}
-    >
-      {isPending ? <Loader size={20} /> : <Trash size={20} />}
-    </button>
+    <div>
+      {/* Bouton de suppression */}
+      <AlertDialog>
+        <AlertDialogTrigger asChild>
+          <Button size="sm" variant="ghost">
+            {isPending ? <Loader size={15} /> : <Trash size={15} />}
+          </Button>
+        </AlertDialogTrigger>
+        <AlertDialogContent>
+          <AlertDialogHeader>
+            <AlertDialogTitle>
+              Êtes-vous sûr de vouloir supprimer ce commentaire ?
+            </AlertDialogTitle>
+            <AlertDialogDescription>
+              Cette action est irréversible.
+            </AlertDialogDescription>
+          </AlertDialogHeader>
+          <AlertDialogFooter>
+            <AlertDialogCancel>Annuler</AlertDialogCancel>
+            <AlertDialogAction
+              onClick={() => {
+                starTransaction(() => {
+                  deleteComment(commentId);
+                });
+              }}
+            >
+              Supprimer
+            </AlertDialogAction>
+          </AlertDialogFooter>
+        </AlertDialogContent>
+      </AlertDialog>
+    </div>
   );
 };
